Rename package to interpreter
