+++ ksqldb-execution/src/main/java/io/confluent/ksql/execution/interpreter/CastInterpreter.java
@@ -13,7 +13,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.ksql.execution.evaluator;
+package io.confluent.ksql.execution.interpreter;
 
 import io.confluent.ksql.execution.codegen.helpers.CastEvaluator;
 import io.confluent.ksql.schema.ksql.SqlBooleans;
