+++ ksqldb-functional-tests/src/test/java/io/confluent/ksql/test/rest/RestQueryTranslationTest.java
@@ -73,7 +73,7 @@ import org.junit.runners.Parameterized;
 @RunWith(Parameterized.class)
 public class RestQueryTranslationTest {
 
-  private static final Path TEST_DIR = Paths.get("rest-query-validation-tests/");
+  private static final Path TEST_DIR = Paths.get("rest-query-validation-tests");
 
   private static final IntegrationTestHarness TEST_HARNESS = IntegrationTestHarness.build();
